componentize the flow run graphs panel
