+++ src/components/index.ts
@@ -91,6 +91,7 @@ export { default as FlowRunCancelModal } from './FlowRunCancelModal.vue'
 export { default as FlowRunCreateForm } from './FlowRunCreateForm.vue'
 export { default as FlowRunDetails } from './FlowRunDetails.vue'
 export { default as FlowRunFilteredList } from './FlowRunFilteredList.vue'
+export { default as FlowRunGraphsPanel } from './FlowRunGraphsPanel.vue'
 export { default as FlowRunIconText } from './FlowRunIconText.vue'
 export { default as FlowRunList } from './FlowRunList.vue'
 export { default as FlowRunListItem } from './FlowRunListItem.vue'
@@ -117,7 +118,6 @@ export { default as FlowRunTaskCount } from './FlowRunTaskCount.vue'
 export { default as FlowRunTaskRuns } from './FlowRunTaskRuns.vue'
 export { default as FlowRunTimeline } from './FlowRunTimeline.vue'
 export { default as FlowRunTimelineOptions } from './FlowRunTimelineOptions.vue'
-export { default as FlowRunTimelineSelectionPanel } from './FlowRunTimelineSelectionPanel.vue'
 export { default as FlowRunTimelineSubFlowRunDetails } from './FlowRunTimelineSubFlowRunDetails.vue'
 export { default as FlowRunTimelineTaskDetails } from './FlowRunTimelineTaskDetails.vue'
 export { default as FlowsDeleteButton } from './FlowsDeleteButton.vue'
